fix: incrementally propagate load tickets and bound chunk cache scopes
