+++ steel-core/src/chunk/simulation_ticket_manager.rs
@@ -1,622 +1,8 @@
-//! Simulation ticket propagation without chunk loading or generation concerns.
-use std::mem;
+//! Simulation propagation stops at the block-ticking boundary.
 
-use rustc_hash::FxHashMap;
-use steel_utils::ChunkPos;
+use super::{chunk_ticket_manager::ChunkTicketLevel, chunk_tracker::ChunkTracker};
 
-use super::{chunk_ticket_manager::ChunkTicketLevel, chunk_ticket_storage::SourceLevelUpdate};
+pub use super::chunk_tracker::ChunkLevelChange as SimulationLevelChange;
 
-const ABSENT_LEVEL: u8 = ChunkTicketLevel::BLOCK_TICKING_CHUNK.raw() + 1;
-const LEVEL_COUNT: u8 = ABSENT_LEVEL + 1;
-
-/// Vanilla-style level buckets for pending graph corrections.
-#[derive(Debug)]
-struct LeveledPropagationQueue {
-    buckets: Vec<Vec<ChunkPos>>,
-    first_queued_level: u8,
-}
-
-impl LeveledPropagationQueue {
-    fn new() -> Self {
-        Self {
-            buckets: (0..usize::from(LEVEL_COUNT)).map(|_| Vec::new()).collect(),
-            first_queued_level: LEVEL_COUNT,
-        }
-    }
-
-    fn enqueue(&mut self, pos: ChunkPos, level: u8) {
-        self.buckets[usize::from(level)].push(pos);
-        self.first_queued_level = self.first_queued_level.min(level);
-    }
-
-    fn pop(&mut self) -> Option<(ChunkPos, u8)> {
-        while self.first_queued_level < LEVEL_COUNT {
-            let bucket = &mut self.buckets[usize::from(self.first_queued_level)];
-            if let Some(pos) = bucket.pop() {
-                return Some((pos, self.first_queued_level));
-            }
-            self.first_queued_level += 1;
-        }
-        None
-    }
-
-    const fn is_empty(&self) -> bool {
-        self.first_queued_level == LEVEL_COUNT
-    }
-}
-
-/// A propagated simulation level change.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct SimulationLevelChange {
-    pub pos: ChunkPos,
-    /// `Some(level)` if the level changed or was added, `None` if it was removed.
-    pub new_level: Option<ChunkTicketLevel>,
-}
-
-/// Tracks simulation sources and their propagated levels.
-#[derive(Debug)]
-pub struct SimulationTicketManager {
-    source_levels: FxHashMap<ChunkPos, u8>,
-    pending_source_levels: FxHashMap<ChunkPos, u8>,
-    levels: FxHashMap<ChunkPos, ChunkTicketLevel>,
-    pending_levels: FxHashMap<ChunkPos, u8>,
-    propagation_queue: LeveledPropagationQueue,
-    changes: Vec<SimulationLevelChange>,
-}
-
-impl Default for SimulationTicketManager {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl SimulationTicketManager {
-    /// Creates an empty simulation ticket manager.
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            source_levels: FxHashMap::default(),
-            pending_source_levels: FxHashMap::default(),
-            levels: FxHashMap::default(),
-            pending_levels: FxHashMap::default(),
-            propagation_queue: LeveledPropagationQueue::new(),
-            changes: Vec::new(),
-        }
-    }
-
-    /// Applies the latest effective simulation source level at one position.
-    pub(crate) fn apply_source_update(&mut self, update: SourceLevelUpdate) {
-        assert!(
-            update.level.is_none_or(ChunkTicketLevel::is_block_ticking),
-            "simulation source level must be block-ticking or stronger"
-        );
-        let new_level = update.level.map_or(ABSENT_LEVEL, ChunkTicketLevel::raw);
-
-        let old_level = self.source_level(update.pos);
-        if old_level == new_level {
-            return;
-        }
-
-        let original_level = *self
-            .pending_source_levels
-            .entry(update.pos)
-            .or_insert(old_level);
-        if new_level == ABSENT_LEVEL {
-            self.source_levels.remove(&update.pos);
-        } else {
-            self.source_levels.insert(update.pos, new_level);
-        }
-
-        if original_level == new_level {
-            self.pending_source_levels.remove(&update.pos);
-        }
-    }
-
-    /// Applies effective source updates in iterator order.
-    pub(crate) fn apply_source_updates(
-        &mut self,
-        updates: impl IntoIterator<Item = SourceLevelUpdate>,
-    ) {
-        for update in updates {
-            self.apply_source_update(update);
-        }
-    }
-
-    /// Applies pending source changes with minimum 8-neighbor propagation.
-    /// The returned changes include additions, updates, and removals.
-    pub fn run_all_updates(&mut self) -> &[SimulationLevelChange] {
-        self.changes.clear();
-
-        if self.pending_source_levels.is_empty() {
-            return &self.changes;
-        }
-
-        debug_assert!(self.pending_levels.is_empty());
-        debug_assert!(self.propagation_queue.is_empty());
-
-        let source_changes = mem::take(&mut self.pending_source_levels);
-        let mut original_levels = FxHashMap::default();
-        for (pos, old_level) in source_changes {
-            let new_level = self.source_level(pos);
-            self.check_edge(None, pos, new_level, new_level < old_level);
-        }
-
-        while let Some((pos, queued_priority)) = self.propagation_queue.pop() {
-            let Some(computed_level) = self.pending_levels.get(&pos).copied() else {
-                continue;
-            };
-            if self.level(pos).min(computed_level) != queued_priority {
-                continue;
-            }
-            self.pending_levels.remove(&pos);
-            self.apply_pending_level(pos, computed_level, &mut original_levels);
-        }
-
-        for (pos, old_level) in original_levels {
-            let new_level = self.levels.get(&pos).copied();
-            if old_level != new_level {
-                self.changes.push(SimulationLevelChange { pos, new_level });
-            }
-        }
-        self.changes
-            .sort_unstable_by_key(|change| (change.pos.0.x, change.pos.0.y));
-
-        &self.changes
-    }
-
-    /// Takes the change buffer produced by the last propagation pass.
-    pub(crate) fn take_changes(&mut self) -> Vec<SimulationLevelChange> {
-        mem::take(&mut self.changes)
-    }
-
-    /// Reuses a drained change buffer on the next propagation pass.
-    pub(crate) fn recycle_changes(&mut self, mut changes: Vec<SimulationLevelChange>) {
-        debug_assert_eq!(self.changes, []);
-        changes.clear();
-        self.changes = changes;
-    }
-
-    /// Returns the last propagated simulation level at `pos`.
-    #[must_use]
-    pub fn get_level(&self, pos: ChunkPos) -> Option<ChunkTicketLevel> {
-        self.levels.get(&pos).copied()
-    }
-
-    fn source_level(&self, pos: ChunkPos) -> u8 {
-        self.source_levels
-            .get(&pos)
-            .copied()
-            .unwrap_or(ABSENT_LEVEL)
-    }
-
-    fn level(&self, pos: ChunkPos) -> u8 {
-        self.levels
-            .get(&pos)
-            .copied()
-            .map_or(ABSENT_LEVEL, ChunkTicketLevel::raw)
-    }
-
-    fn apply_pending_level(
-        &mut self,
-        pos: ChunkPos,
-        computed_level: u8,
-        original_levels: &mut FxHashMap<ChunkPos, Option<ChunkTicketLevel>>,
-    ) {
-        let current_level = self.level(pos);
-        if computed_level < current_level {
-            self.set_level(pos, computed_level, original_levels);
-            self.check_neighbors_after_update(pos, computed_level, true);
-        } else if computed_level > current_level {
-            self.set_level(pos, ABSENT_LEVEL, original_levels);
-            if computed_level != ABSENT_LEVEL {
-                self.schedule_level(pos, computed_level);
-            }
-            self.check_neighbors_after_update(pos, current_level, false);
-        }
-    }
-
-    fn set_level(
-        &mut self,
-        pos: ChunkPos,
-        level: u8,
-        original_levels: &mut FxHashMap<ChunkPos, Option<ChunkTicketLevel>>,
-    ) {
-        original_levels
-            .entry(pos)
-            .or_insert_with(|| self.levels.get(&pos).copied());
-
-        if level == ABSENT_LEVEL {
-            self.levels.remove(&pos);
-            return;
-        }
-
-        let Some(level) = ChunkTicketLevel::new(level) else {
-            panic!("propagated simulation level exceeds ChunkTicketLevel::MAX");
-        };
-        self.levels.insert(pos, level);
-    }
-
-    fn check_neighbors_after_update(&mut self, pos: ChunkPos, level: u8, only_decrease: bool) {
-        if only_decrease && level >= ChunkTicketLevel::BLOCK_TICKING_CHUNK.raw() {
-            return;
-        }
-
-        for neighbor in pos.neighbors() {
-            self.check_neighbor(pos, neighbor, level, only_decrease);
-        }
-    }
-
-    fn check_neighbor(
-        &mut self,
-        from: ChunkPos,
-        to: ChunkPos,
-        from_level: u8,
-        only_decrease: bool,
-    ) {
-        let propagated_level = from_level.saturating_add(1).min(ABSENT_LEVEL);
-        if only_decrease {
-            self.check_edge(Some(from), to, propagated_level, true);
-            return;
-        }
-
-        let old_computed_level = self
-            .pending_levels
-            .get(&to)
-            .copied()
-            .unwrap_or_else(|| self.level(to));
-        if propagated_level == old_computed_level {
-            self.check_edge(Some(from), to, ABSENT_LEVEL, false);
-        }
-    }
-
-    fn check_edge(
-        &mut self,
-        known_parent: Option<ChunkPos>,
-        pos: ChunkPos,
-        level_from_parent: u8,
-        only_decrease: bool,
-    ) {
-        let current_level = self.level(pos);
-        let old_computed_level = self
-            .pending_levels
-            .get(&pos)
-            .copied()
-            .unwrap_or(current_level);
-        let new_computed_level = if only_decrease {
-            old_computed_level.min(level_from_parent)
-        } else {
-            self.compute_level(pos, known_parent, level_from_parent)
-        };
-
-        self.replace_pending_level(pos, current_level, new_computed_level);
-    }
-
-    fn compute_level(
-        &self,
-        pos: ChunkPos,
-        known_parent: Option<ChunkPos>,
-        level_from_parent: u8,
-    ) -> u8 {
-        let mut computed_level = level_from_parent.min(self.source_level(pos));
-        for neighbor in pos.neighbors() {
-            if Some(neighbor) == known_parent {
-                continue;
-            }
-
-            computed_level = computed_level.min(self.level(neighbor).saturating_add(1));
-            if computed_level == 0 {
-                break;
-            }
-        }
-        computed_level.min(ABSENT_LEVEL)
-    }
-
-    fn replace_pending_level(&mut self, pos: ChunkPos, current_level: u8, computed_level: u8) {
-        if current_level == computed_level {
-            self.pending_levels.remove(&pos);
-        } else {
-            self.pending_levels.insert(pos, computed_level);
-            self.propagation_queue
-                .enqueue(pos, current_level.min(computed_level));
-        }
-    }
-
-    fn schedule_level(&mut self, pos: ChunkPos, computed_level: u8) {
-        let current_level = self.level(pos);
-        debug_assert!(!self.pending_levels.contains_key(&pos));
-        self.pending_levels.insert(pos, computed_level);
-        self.propagation_queue
-            .enqueue(pos, current_level.min(computed_level));
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    const REFERENCE_BLOCK_TICKING_LEVEL: u8 = ChunkTicketLevel::BLOCK_TICKING_CHUNK.raw();
-    const DETERMINISTIC_SEQUENCE_MULTIPLIER: u64 = 6_364_136_223_846_793_005;
-    const DETERMINISTIC_SEQUENCE_INCREMENT: u64 = 1_442_695_040_888_963_407;
-
-    struct DeterministicSequence(u64);
-
-    impl DeterministicSequence {
-        fn next(&mut self, bound: usize) -> usize {
-            self.0 = self
-                .0
-                .wrapping_mul(DETERMINISTIC_SEQUENCE_MULTIPLIER)
-                .wrapping_add(DETERMINISTIC_SEQUENCE_INCREMENT);
-            let bound = u64::try_from(bound).expect("test bound must fit in u64");
-            let value = (self.0 >> u32::BITS) % bound;
-            usize::try_from(value).expect("bounded test value must fit in usize")
-        }
-    }
-
-    const fn source(pos: ChunkPos, level: Option<ChunkTicketLevel>) -> SourceLevelUpdate {
-        SourceLevelUpdate { pos, level }
-    }
-
-    const fn entity_ticking_source(pos: ChunkPos, radius: u8) -> SourceLevelUpdate {
-        source(
-            pos,
-            Some(ChunkTicketLevel::for_entity_ticking_radius(radius)),
-        )
-    }
-
-    fn has_change(
-        changes: &[SimulationLevelChange],
-        pos: ChunkPos,
-        new_level: Option<ChunkTicketLevel>,
-    ) -> bool {
-        changes.contains(&SimulationLevelChange { pos, new_level })
-    }
-
-    fn add_reference_source(
-        levels: &mut FxHashMap<ChunkPos, ChunkTicketLevel>,
-        source_pos: ChunkPos,
-        source_level: u8,
-    ) {
-        if source_level > REFERENCE_BLOCK_TICKING_LEVEL {
-            return;
-        }
-
-        let radius = i32::from(REFERENCE_BLOCK_TICKING_LEVEL - source_level);
-        for dz in -radius..=radius {
-            for dx in -radius..=radius {
-                let distance = u8::try_from(dx.unsigned_abs().max(dz.unsigned_abs()))
-                    .expect("reference distance must fit in u8");
-                let raw_level = source_level
-                    .checked_add(distance)
-                    .expect("reference level must fit in u8");
-                let Some(level) = ChunkTicketLevel::new(raw_level) else {
-                    panic!("reference produced an invalid simulation level");
-                };
-                let pos = ChunkPos::new(source_pos.0.x + dx, source_pos.0.y + dz);
-                levels
-                    .entry(pos)
-                    .and_modify(|stored| *stored = (*stored).min(level))
-                    .or_insert(level);
-            }
-        }
-    }
-
-    fn reference_levels(
-        manager: &SimulationTicketManager,
-    ) -> FxHashMap<ChunkPos, ChunkTicketLevel> {
-        let mut levels = FxHashMap::default();
-        for (&pos, &source_level) in &manager.source_levels {
-            add_reference_source(&mut levels, pos, source_level);
-        }
-        levels
-    }
-
-    fn reference_changes(
-        old_levels: &FxHashMap<ChunkPos, ChunkTicketLevel>,
-        new_levels: &FxHashMap<ChunkPos, ChunkTicketLevel>,
-    ) -> Vec<SimulationLevelChange> {
-        let mut changes = Vec::new();
-        for (&pos, &new_level) in new_levels {
-            if old_levels.get(&pos) != Some(&new_level) {
-                changes.push(SimulationLevelChange {
-                    pos,
-                    new_level: Some(new_level),
-                });
-            }
-        }
-        for &pos in old_levels.keys() {
-            if !new_levels.contains_key(&pos) {
-                changes.push(SimulationLevelChange {
-                    pos,
-                    new_level: None,
-                });
-            }
-        }
-        changes.sort_unstable_by_key(|change| (change.pos.0.x, change.pos.0.y));
-        changes
-    }
-
-    fn run_and_compare_with_reference(
-        manager: &mut SimulationTicketManager,
-        previous_levels: &mut FxHashMap<ChunkPos, ChunkTicketLevel>,
-    ) {
-        let actual_changes = manager.run_all_updates().to_vec();
-        let expected_levels = reference_levels(manager);
-        let expected_changes = reference_changes(previous_levels, &expected_levels);
-
-        assert_eq!(manager.levels, expected_levels);
-        assert_eq!(actual_changes, expected_changes);
-        *previous_levels = expected_levels;
-    }
-
-    #[test]
-    fn overlapping_sources_keep_the_strongest_propagated_level() {
-        let mut manager = SimulationTicketManager::new();
-        manager.apply_source_updates([
-            entity_ticking_source(ChunkPos::new(0, 0), 2),
-            entity_ticking_source(ChunkPos::new(4, 0), 0),
-        ]);
-        manager.run_all_updates();
-
-        assert_eq!(
-            manager.get_level(ChunkPos::new(2, 0)),
-            Some(ChunkTicketLevel::ENTITY_TICKING_CHUNK)
-        );
-        assert_eq!(
-            manager.get_level(ChunkPos::new(3, 0)),
-            Some(ChunkTicketLevel::BLOCK_TICKING_CHUNK)
-        );
-        assert_eq!(
-            manager.get_level(ChunkPos::new(4, 0)),
-            Some(ChunkTicketLevel::ENTITY_TICKING_CHUNK)
-        );
-    }
-
-    #[test]
-    fn repeated_updates_coalesce_to_the_original_level() {
-        let mut manager = SimulationTicketManager::new();
-        let pos = ChunkPos::new(0, 0);
-        manager.apply_source_update(entity_ticking_source(pos, 2));
-        manager.apply_source_update(entity_ticking_source(pos, 0));
-        manager.apply_source_update(source(pos, None));
-
-        assert_eq!(manager.run_all_updates(), []);
-        assert_eq!(manager.run_all_updates(), []);
-        assert_eq!(manager.get_level(pos), None);
-
-        manager.apply_source_updates([source(pos, None), entity_ticking_source(pos, 2)]);
-        assert_ne!(manager.run_all_updates(), []);
-        manager.apply_source_update(entity_ticking_source(pos, 0));
-        manager.apply_source_update(entity_ticking_source(pos, 2));
-
-        assert_eq!(manager.run_all_updates(), []);
-        assert_eq!(
-            manager.get_level(pos),
-            Some(ChunkTicketLevel::for_entity_ticking_radius(2))
-        );
-    }
-
-    #[test]
-    #[should_panic(expected = "simulation source level must be block-ticking or stronger")]
-    fn load_only_level_cannot_alias_an_absent_simulation_source() {
-        let mut manager = SimulationTicketManager::new();
-        manager.apply_source_update(SourceLevelUpdate {
-            pos: ChunkPos::new(0, 0),
-            level: Some(ChunkTicketLevel::FULL_CHUNK),
-        });
-    }
-
-    #[test]
-    fn weakening_a_source_removes_its_old_outer_levels() {
-        let mut manager = SimulationTicketManager::new();
-        let pos = ChunkPos::new(0, 0);
-        manager.apply_source_update(entity_ticking_source(pos, 2));
-        manager.run_all_updates();
-
-        manager.apply_source_update(entity_ticking_source(pos, 0));
-        let changes = manager.run_all_updates();
-
-        assert!(has_change(
-            changes,
-            pos,
-            Some(ChunkTicketLevel::ENTITY_TICKING_CHUNK)
-        ));
-        assert!(has_change(changes, ChunkPos::new(2, 0), None));
-        assert_eq!(
-            manager.get_level(pos),
-            Some(ChunkTicketLevel::ENTITY_TICKING_CHUNK)
-        );
-    }
-
-    #[test]
-    fn source_batch_order_does_not_change_the_result() {
-        let positions = [
-            ChunkPos::new(0, 0),
-            ChunkPos::new(6, 1),
-            ChunkPos::new(-3, 2),
-            ChunkPos::new(4, -2),
-        ];
-        let levels = [
-            Some(ChunkTicketLevel::for_entity_ticking_radius(4)),
-            Some(ChunkTicketLevel::for_entity_ticking_radius(1)),
-            Some(ChunkTicketLevel::for_entity_ticking_radius(2)),
-            Some(ChunkTicketLevel::for_entity_ticking_radius(0)),
-        ];
-        let mut forwards = SimulationTicketManager::new();
-        let mut backwards = SimulationTicketManager::new();
-
-        forwards.apply_source_updates(
-            positions
-                .into_iter()
-                .zip(levels)
-                .map(|(pos, level)| source(pos, level)),
-        );
-        backwards.apply_source_updates(
-            positions
-                .into_iter()
-                .zip(levels)
-                .rev()
-                .map(|(pos, level)| source(pos, level)),
-        );
-
-        assert_eq!(forwards.run_all_updates(), backwards.run_all_updates());
-        assert_eq!(forwards.levels, backwards.levels);
-    }
-
-    #[test]
-    fn incremental_source_updates_match_reference() {
-        let mut manager = SimulationTicketManager::new();
-        let mut previous_levels = FxHashMap::default();
-        manager.apply_source_updates([
-            entity_ticking_source(ChunkPos::new(0, 0), 4),
-            entity_ticking_source(ChunkPos::new(6, 1), 1),
-            entity_ticking_source(ChunkPos::new(-3, 2), 2),
-            entity_ticking_source(ChunkPos::new(4, -2), 0),
-        ]);
-        run_and_compare_with_reference(&mut manager, &mut previous_levels);
-
-        for level in [
-            Some(ChunkTicketLevel::for_entity_ticking_radius(0)),
-            Some(ChunkTicketLevel::for_entity_ticking_radius(1)),
-            Some(ChunkTicketLevel::for_entity_ticking_radius(3)),
-            Some(ChunkTicketLevel::BLOCK_TICKING_CHUNK),
-        ] {
-            manager.apply_source_update(source(ChunkPos::new(-3, 2), level));
-            run_and_compare_with_reference(&mut manager, &mut previous_levels);
-        }
-
-        manager.apply_source_updates([
-            source(ChunkPos::new(0, 0), None),
-            source(ChunkPos::new(-3, 2), None),
-            entity_ticking_source(ChunkPos::new(9, 3), 3),
-            source(ChunkPos::new(4, -2), None),
-            source(ChunkPos::new(6, 1), None),
-        ]);
-        run_and_compare_with_reference(&mut manager, &mut previous_levels);
-    }
-
-    #[test]
-    fn deterministic_random_operations_match_reference() {
-        let mut manager = SimulationTicketManager::new();
-        let mut previous_levels = FxHashMap::default();
-        let mut sequence = DeterministicSequence(0x5eed_cafe_d00d_f00d);
-        let source_levels = [
-            None,
-            Some(ChunkTicketLevel::for_entity_ticking_radius(4)),
-            Some(ChunkTicketLevel::for_entity_ticking_radius(2)),
-            Some(ChunkTicketLevel::for_entity_ticking_radius(0)),
-            Some(ChunkTicketLevel::BLOCK_TICKING_CHUNK),
-        ];
-
-        for _ in 0..200 {
-            for _ in 0..=sequence.next(3) {
-                let x = i32::try_from(sequence.next(13)).expect("test x must fit in i32") - 6;
-                let z = i32::try_from(sequence.next(13)).expect("test z must fit in i32") - 6;
-                let pos = ChunkPos::new(x, z);
-                let level = source_levels[sequence.next(source_levels.len())];
-                manager.apply_source_update(source(pos, level));
-            }
-
-            run_and_compare_with_reference(&mut manager, &mut previous_levels);
-        }
-    }
-}
+/// Vanilla's simulation tracker, with an independent graph and ticking limit.
+pub type SimulationTicketManager = ChunkTracker<{ ChunkTicketLevel::BLOCK_TICKING_CHUNK.raw() }>;
